+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@moondreamsdev/dreamer-ui",
-  "version": "1.7.10-test.52",
+  "version": "1.7.10-test.53",
   "publishConfig": {
     "access": "public"
   },
