+++ lib/src/components/popover/index.ts
@@ -1,0 +1,1 @@
+export { Popover, type PopoverAlignment, type PopoverProps } from './Popover';
