+++ lib/src/utils/index.ts
@@ -1,2 +1,3 @@
 export { downloadFile } from './downloadFile';
 export { join } from './join';
+export { mergeRefs } from './mergeRefs';
