+++ lib/src/utils/mergeRefs.ts
